refactor OrgCatalog
