+++ frontend/src/modules/OrgCatalogue/hooks/useSearch.js
@@ -1,0 +1,14 @@
+import { useState } from 'react';
+
+export const useSearch = (initialValue) => {
+  const [searchValue, setSearchValue] = useState(initialValue || '');
+
+  const handleSearchChange = (e) => {
+    setSearchValue(e.target.value);
+  };
+
+  return {
+    searchValue,
+    handleSearchChange,
+  };
+};
